+++ MaternityHospital/Services/PregnancyCalculator.cs
@@ -8,18 +8,18 @@ namespace MaternityHospital.Services
 {
     public class PregnancyCalculator : IPregnancyCalculator
     {
-        protected double PregnancyDuration { get; }
+        protected int PregnancyDuration { get; }
         protected DateTime? LastPeriodDate { get; }
         protected DateTime FirstScanDate { get; }
 
-        public PregnancyCalculator(DateTime? lastPeriodDate, DateTime firstScanDate, double pregnancyDuration = 0)
+        public PregnancyCalculator(DateTime? lastPeriodDate, DateTime firstScanDate, int pregnancyDuration = 0)
         {
             LastPeriodDate = lastPeriodDate;
             FirstScanDate = firstScanDate;
             PregnancyDuration = pregnancyDuration;
         }
 
-        public double Calculate()
+        public int Calculate()
         {
             if(LastPeriodDate == null)
             {
@@ -27,11 +27,11 @@ namespace MaternityHospital.Services
             }
             return PeriodCalculate();
         }
-        private double ScanCalculate()
+        private int ScanCalculate()
         {
             return (DateTime.Now - FirstScanDate).Days/7 + PregnancyDuration;
         }
-        private double PeriodCalculate()
+        private int PeriodCalculate()
         {
             return (DateTime.Now - LastPeriodDate).Value.Days/7 + PregnancyDuration;
         }
